+++ www/utils/packages/docs-automator/src/utils/dirname.ts
@@ -1,0 +1,8 @@
+import path from "path"
+import { fileURLToPath } from "url"
+
+export default function dirname(fileUrl: string) {
+  const __filename = fileURLToPath(fileUrl)
+
+  return path.dirname(__filename)
+}