+++ packages/design-system/ui/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @medusajs/ui
 
+## 4.1.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/icons@2.13.6
+
+## 4.1.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/icons@2.13.5
+
 ## 4.1.4
 
 ### Patch Changes